+++ service/service.go
@@ -58,6 +58,7 @@ type Config struct {
 	ArangodPath          string
 	ArangodJSPath        string
 	ArangoSyncPath       string
+	AdvertisedEndpoint   string
 	MasterPort           int
 	RrPath               string
 	DataDir              string